requirements.txt
osmnx 
requests
shapely 
trimesh 
numpy-stl